chat is working!!
Register, Login and chat is working on localhost. The problem was that I was creating the const db using the sqlite package, which provides a Promise-based API, so db could only support async/await methods. I was using db with callback methods, which don't work with sqlite APIs. Instead, they return a Promise that must be handled with await or .then().